refactor(MOR-610): D4 move internal audio callers to canonical import paths
Internal callers come off the sys.modules-aliased back-compat shims
(shims themselves untouched):

- audio/bridge.py: rigplane.core.radio_protocol.AudioCapable,
  rigplane.core.types.AudioCodec
- audio/lan_stream.py: rigplane.core.transport.IcomTransport,
  rigplane.core.types.PacketType
- web/handlers/audio.py: rigplane.audio._codecs / rigplane.audio._transcoder
- runtime/radio.py: rigplane.audio.bus.AudioBus
- web/server.py + cli/__init__.py: rigplane.audio.bridge.*

lint-imports: 4 kept / 0 broken.
